move most of logic from certificatebutton to custom hook
