+++ frontend/components/CertificateButton.tsx
@@ -1,6 +1,5 @@
-import { useContext, useEffect, useReducer, useState } from "react"
+import { useCallback, useContext, useState } from "react"
 
-import { useMutation } from "@apollo/client"
 import styled from "@emotion/styled"
 import {
   CircularProgress,
@@ -14,19 +13,11 @@ import DialogContent from "@mui/material/DialogContent"
 import DialogTitle from "@mui/material/DialogTitle"
 
 import AlertContext from "/contexts/AlertContext"
-import LoginStateContext from "/contexts/LoginStateContext"
-import { updateAccount } from "/lib/account"
-import { checkCertificate, createCertificate } from "/lib/certificates"
+import { useCertificate } from "/hooks/useCertificate"
 import CompletionsTranslations from "/translations/completions"
 import { useTranslator } from "/util/useTranslator"
 
-import {
-  CourseCoreFieldsFragment,
-  CurrentUserDetailedDocument,
-  CurrentUserDocument,
-  CurrentUserOverviewDocument,
-  UpdateUserNameDocument,
-} from "/graphql/generated"
+import { CourseCoreFieldsFragment } from "/graphql/generated"
 
 const StyledButton = styled(Button)`
   margin: auto;
@@ -44,249 +35,115 @@ interface CertificateProps {
   course: CourseCoreFieldsFragment
 }
 
-type Status =
-  | "IDLE"
-  | "CHECKING"
-  | "AVAILABLE"
-  | "NOT_AVAILABLE"
-  | "GENERATING"
-  | "UPDATING_NAME"
-  | "ERROR"
-type ActionType =
-  | "RESET"
-  | "CHECK_CERTIFICATE"
-  | "RECEIVE_CERTIFICATE_ID"
-  | "GENERATE_CERTIFICATE"
-  | "RECEIVE_GENERATED_CERTIFICATE"
-  | "UPDATE_NAME"
-  | "UPDATED_NAME"
-  | "ERROR"
-
-type Action = {
-  type: ActionType
-  payload?: any
-}
-
-interface CertificateState {
-  status: Status
-  certificateId?: string
-  error?: string
-}
-
-const initialState: CertificateState = {
-  status: "IDLE",
-}
-
-const reducer = (state: CertificateState, action: Action): CertificateState => {
-  switch (action.type) {
-    case "RESET":
-      return initialState
-    case "CHECK_CERTIFICATE":
-      return {
-        ...state,
-        status: "CHECKING",
-      }
-    case "RECEIVE_CERTIFICATE_ID":
-      return {
-        ...state,
-        status: "IDLE",
-        certificateId: action?.payload,
-        error: undefined,
-      }
-    case "GENERATE_CERTIFICATE":
-      return {
-        ...state,
-        status: "GENERATING",
-        certificateId: undefined,
-        error: undefined,
-      }
-    case "RECEIVE_GENERATED_CERTIFICATE":
-      return {
-        ...state,
-        status: "IDLE",
-        certificateId: action?.payload?.id,
-        error: undefined,
-      }
-    case "UPDATE_NAME":
-      return {
-        ...state,
-        status: "UPDATING_NAME",
-        error: undefined,
-      }
-    case "UPDATED_NAME":
-      return {
-        ...state,
-        status: "IDLE",
-        error: undefined,
-      }
-    case "ERROR":
-      return {
-        ...state,
-        status: "ERROR",
-        error: action.payload,
-      }
-    default:
-      return state
-  }
-}
-
 const CertificateButton = ({ course }: CertificateProps) => {
   const t = useTranslator(CompletionsTranslations)
-  const { currentUser, updateUser, admin } = useContext(LoginStateContext)
   const { addAlert } = useContext(AlertContext)
+  const [dialogOpen, setDialogOpen] = useState(false)
 
-  const [state, dispatch] = useReducer(reducer, initialState)
-  const [firstName, setFirstName] = useState(currentUser?.first_name ?? "")
-  const [lastName, setLastName] = useState(currentUser?.last_name ?? "")
-
-  const [updateUserName] = useMutation(UpdateUserNameDocument, {
-    refetchQueries: [
-      { query: CurrentUserDocument },
-      { query: CurrentUserDetailedDocument },
-      { query: CurrentUserOverviewDocument },
-    ],
-  })
-
-  const [open, setOpen] = useState(false)
-
-  useEffect(() => {
-    dispatch({ type: "CHECK_CERTIFICATE" })
-    checkCertificate(course.slug)
-      .then((res: any) => {
-        dispatch({
-          type: "RECEIVE_CERTIFICATE_ID",
-          payload: res?.existing_certificate,
-        })
-      })
-      .catch((e: any) => {
-        dispatch({ type: "ERROR", payload: e })
-        console.error("error?", e)
-        addAlert({
-          title: t("nameFormErrorTitle"),
-          message: t("nameFormErrorCheckCertificate"),
-          severity: "error",
-        })
-        dispatch({ type: "RESET" })
-      })
+  const onCertificateCheckSuccess = useCallback(() => {
+    addAlert({
+      title: t("nameFormErrorTitle"),
+      message: t("nameFormErrorCheckCertificate"),
+      severity: "error",
+    })
+  }, [])
+  const onReceiveGeneratedCertificateSuccess = useCallback(() => {
+    setDialogOpen(false)
+  }, [])
+  const onReceiveGeneratedCertificateError = useCallback(() => {
+    setDialogOpen(false)
+    addAlert({
+      title: t("nameFormErrorTitle"),
+      message: t("nameFormErrorSubmit"),
+      severity: "error",
+    })
   }, [])
 
-  // TODO: when admin is looking at a user, it's checking against the admin's name -- fix
+  const {
+    state,
+    firstName,
+    setFirstName,
+    lastName,
+    setLastName,
+    onSubmit,
+    isNameEmpty,
+  } = useCertificate({
+    course,
+    onCertificateCheckSuccess,
+    onReceiveGeneratedCertificateSuccess,
+    onReceiveGeneratedCertificateError,
+  })
 
-  const nameChanged = () =>
-    (currentUser?.first_name ?? "") !== firstName ||
-    (currentUser?.last_name ?? "") !== lastName
-
-  const isNameEmpty = () => firstName == "" && lastName == ""
-
-  const onSubmit = async () => {
-    if (isNameEmpty()) {
-      return
-    }
-
-    try {
-      if (nameChanged()) {
-        dispatch({ type: "UPDATE_NAME" })
-        const res = await updateAccount(firstName, lastName)
-        await updateUserName({
-          variables: {
-            first_name: firstName,
-            last_name: lastName,
-          },
-        })
-        updateUser({
-          user: { ...currentUser!, first_name: firstName, last_name: lastName },
-          admin,
-        })
-        dispatch({ type: "UPDATED_NAME", payload: res })
-      }
-
-      dispatch({ type: "GENERATE_CERTIFICATE" })
-      const res = await createCertificate(course.slug)
-      dispatch({ type: "RECEIVE_GENERATED_CERTIFICATE", payload: res })
-      setOpen(false)
-    } catch (e) {
-      dispatch({ type: "ERROR", payload: e })
-      setOpen(false)
-      console.error("error?", e)
-      addAlert({
-        title: t("nameFormErrorTitle"),
-        message: t("nameFormErrorSubmit"),
-        severity: "error",
-      })
-      dispatch({ type: "RESET" })
-    }
+  if (state.certificateId) {
+    return (
+      <StyledButton
+        onClick={() =>
+          window.open(
+            `https://certificates.mooc.fi/validate/${state.certificateId}`,
+            "_blank",
+          )
+        }
+      >
+        {t("showCertificate")}
+      </StyledButton>
+    )
   }
 
   return (
     <>
-      {state.certificateId ? (
-        <StyledButton
-          onClick={() =>
-            window.open(
-              `https://certificates.mooc.fi/validate/${state.certificateId}`,
-              "_blank",
-            )
-          }
-        >
-          {t("showCertificate")}
-        </StyledButton>
-      ) : (
-        <>
-          <StyledButton onClick={() => setOpen(true)}>
-            {t("createCertificate")}
-          </StyledButton>
-          <StyledDialog
-            open={open}
-            onClose={() => setOpen(false)}
-            aria-labelledby="dialog-title"
+      <StyledButton onClick={() => setDialogOpen(true)}>
+        {t("createCertificate")}
+      </StyledButton>
+      <StyledDialog
+        open={dialogOpen}
+        onClose={() => setDialogOpen(false)}
+        aria-labelledby="dialog-title"
+      >
+        <DialogTitle id="dialog-title">{t("nameFormTitle")}</DialogTitle>
+        <DialogContent>
+          <DialogContentText style={{ marginBottom: "1.5rem" }}>
+            {t("nameFormIntro")}
+          </DialogContentText>
+          <StyledTextField
+            id="first-name"
+            label={t("nameFormFirstName")}
+            value={firstName}
+            onChange={(event) => setFirstName(event.target.value)}
+            defaultValue={firstName}
+            fullWidth
+          />
+          <StyledTextField
+            id="last-name"
+            label={t("nameFormLastName")}
+            defaultValue={lastName}
+            value={lastName}
+            onChange={(event) => setLastName(event.target.value)}
+            fullWidth
+          />
+        </DialogContent>
+        <DialogActions>
+          <Button
+            onClick={() => setDialogOpen(false)}
+            variant="outlined"
+            color="inherit"
+            fullWidth
           >
-            <DialogTitle id="dialog-title">{t("nameFormTitle")}</DialogTitle>
-            <DialogContent>
-              <DialogContentText style={{ marginBottom: "1.5rem" }}>
-                {t("nameFormIntro")}
-              </DialogContentText>
-              <StyledTextField
-                id="first-name"
-                label={t("nameFormFirstName")}
-                value={firstName}
-                onChange={(event) => setFirstName(event.target.value)}
-                defaultValue={firstName}
-                fullWidth
-              />
-              <StyledTextField
-                id="last-name"
-                label={t("nameFormLastName")}
-                defaultValue={lastName}
-                value={lastName}
-                onChange={(event) => setLastName(event.target.value)}
-                fullWidth
-              />
-            </DialogContent>
-            <DialogActions>
-              <Button
-                onClick={() => setOpen(false)}
-                variant="outlined"
-                color="inherit"
-                fullWidth
-              >
-                {t("nameFormCancel")}
-              </Button>
-              <Button
-                onClick={onSubmit}
-                disabled={state.status !== "IDLE" || isNameEmpty()}
-                color="primary"
-                fullWidth
-              >
-                {["IDLE", "ERROR"].indexOf(state.status) < 0 ? (
-                  <CircularProgress size={24} color="secondary" />
-                ) : (
-                  t("nameFormChangeAndSubmit")
-                )}
-              </Button>
-            </DialogActions>
-          </StyledDialog>
-        </>
-      )}
+            {t("nameFormCancel")}
+          </Button>
+          <Button
+            onClick={onSubmit}
+            disabled={state.status !== "IDLE" || isNameEmpty}
+            color="primary"
+            fullWidth
+          >
+            {["IDLE", "ERROR"].indexOf(state.status) < 0 ? (
+              <CircularProgress size={24} color="secondary" />
+            ) : (
+              t("nameFormChangeAndSubmit")
+            )}
+          </Button>
+        </DialogActions>
+      </StyledDialog>
     </>
   )
 }
